// cal-to-json/src/cal/property-type.ts
export enum PropertyType {
  Text = 'TEXT',
  Boolean = 'BOOLEAN',
  TextML = 'TEXT_ML',
  TableRelation = 'TABLE_RELATION',
  Integer = 'INTEGER',
  FieldList = 'FIELD_LIST',
  Trigger = 'TRIGGER',
  CalcFormula = 'CALC_FORMULA',
  Permission = 'PERMISSION',
  Permissions = 'PERMISSIONS',
  Option = 'OPTION',
  TableView = 'TABLE_VIEW',
  TableFilter = 'TABLE_FILTER',
  DataItemLink = 'DATA_ITEM_LINK',
  ActionList = 'ACTION_LIST',
  OrderBy = 'ORDERBY',
}